Added timer to wash money
Added a timer to wash the money. The player does not have to stay at the wash for the timer to complete. Also cleaned up some of the menu code. Had a section throwing a small error.

Timer settings can be enabled or disabled in config file and the length of time can also be set in the config file.